# ChasmaWebApi/ChasmaWebOpenApiSpec.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Chasma Git Manager API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:5001"
    }
  ],
  "paths": {
    "/api/Branch/addBranch": {
      "post": {
        "tags": [
          "Branch"
        ],
        "operationId": "Branch_AddNewBranch",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNewBranchRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNewBranchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Branch/deleteBranch": {
      "delete": {
        "tags": [
          "Branch"
        ],
        "operationId": "Branch_DeleteBranch",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteBranchRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteBranchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Branch/gitCheckout": {
      "post": {
        "tags": [
          "Branch"
        ],
        "operationId": "Branch_CheckoutBranch",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitCheckoutRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitCheckoutResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Branch/gitBranch": {
      "post": {
        "tags": [
          "Branch"
        ],
        "operationId": "Branch_GetBranches",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitBranchRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitBranchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Branch/gitMerge": {
      "post": {
        "tags": [
          "Branch"
        ],
        "operationId": "Branch_MergeBranch",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitMergeRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitMergeResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddNewBranchResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ResponseBase"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "ResponseBase": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ChasmaXmlBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "isErrorResponse": {
                "type": "boolean"
              },
              "errorMessage": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "ChasmaXmlBase": {
        "type": "object",
        "additionalProperties": false
      },
      "AddNewBranchRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ChasmaXmlBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "repositoryId": {
                "type": "string"
              },
              "userId": {
                "type": "integer",
                "format": "int32"
              },
              "branchName": {
                "type": "string"
              },
              "isCheckingOutNewBranch": {
                "type": "boolean"
              }
            }
          }
        ]
      },
      "DeleteBranchResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ResponseBase"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "DeleteBranchRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ChasmaXmlBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "repositoryId": {
                "type": "string"
              },
              "branchName": {
                "type": "string"
              }
            }
          }
        ]
      },
      "GitCheckoutResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ResponseBase"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "GitCheckoutRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ChasmaXmlBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "repositoryId": {
                "type": "string"
              },
              "branchName": {
                "type": "string"
              }
            }
          }
        ]
      },
      "GitBranchResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ResponseBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "branchNames": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "GitBranchRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ChasmaXmlBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "repositoryId": {
                "type": "string"
              }
            }
          }
        ]
      },
      "GitMergeResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ResponseBase"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "GitMergeRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ChasmaXmlBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "repositoryId": {
                "type": "string"
              },
              "userId": {
                "type": "integer",
                "format": "int32"
              },
              "sourceBranch": {
                "type": "string"
              },
              "destinationBranch": {
                "type": "string"
              }
            }
          }
        ]
      },
...
    }
  }
}